Update LittleTilesProfilerOverlay.java
I spent a good bit of time pulling out my hair trying to figure out which mod added the message about mipmap levels. It wasn't obvious, and I think this edit should at least make it clearer who is printing them and at least where to look in the case of turning that one off in the config.